Extract pool-restrictions
