+++ src/composables/usePoolHelpers.ts
@@ -540,17 +540,6 @@ export function fiatValueOf(pool: Pool, shares: string): string {
 }
 
 /**
- * Checks if pool ID is included in the list of pools that joins should be
- * disabled for, e.g. you can't access the invest page.
- *
- * @param {string} id - The pool ID to check
- * @returns {boolean} True if included in list.
- */
-export function isJoinsDisabled(id: string): boolean {
-  return POOLS.DisabledJoins.includes(id.toLowerCase());
-}
-
-/**
  * Checks if pool ID is included in the list of deprecated pools
  * @param {string} id - The pool ID to check
  * @returns {boolean} True if included in list
